+++ CPPProject2/pi_approx.cpp
@@ -1,17 +1,19 @@
 #include <math.h>
 
 struct PiResults {
-        float approx;
-        float error;
+        double approx;
+        double error;
     };
 
 PiResults pi_approx(int N) {
     PiResults results;
-    float sum = 0;
-    float dx = 1.0/float(N);
+    double sum = 0.0;
+    double dx = 1.0/double(N);
     
-    for (float k=1; k<=N; k++) {
-        sum += dx*(sqrt(1-pow(k/float(N),2))+sqrt(1-pow((k-1)/float(N),2)))/2;
+    for (int k=1; k<=N; k++) {
+        sum += dx*(sqrt(1.0-pow(double(k)/double(N),2.0))+sqrt(1.0-pow(double(k-1)/double(N),2.0)))/2.0;
+        
+            
     }
     results.approx = sum*4.0;
